debt_manager_backend_fastapi/requirements.txt

fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.6

sqlalchemy
